requirements.txt
crawl4AI ~= 0.4.247
asyncio ~= 3.4.3
sentence-transformers ~=3.4.1
langchain_ollama ~=0.2.3
fastapi[standard]~=0.115.8
typing ~=3.7.4.3
python-dotenv~=1.0.1
langchain_chroma ~=0.2.1
langchain_text_splitters ~=0.3.6
loguru~=0.7.3
praw~=7.8.1
pandas~=2.2.3
textblob~= 0.19.0
psaw~=0.1.0
nltk~=3.9.1
gensim~=4.3.3
sumy ~=0.11.0
pytesseract~=0.3.13
opencv-python~=4.11.0.86
numpy~=1.26.4
dask~=2025.2.0
distributed~=2025.02.0

